Adds support for binary concurrency check columns.
 They should be defined as;

 [Type(Size = 8)]
 [Column(Flags = ColumnFlags.Generated | ColumnFlags.ConcurrencyCheck)]
 public byte[] Version { get; set; }

 The size can be 2, 4 or 8.

 SQLite does not support this feature.

+++ FoxDb.Core/Config/TypeFactory.cs
@@ -20,7 +20,7 @@ namespace FoxDb
             }
         }
 
-        public ITypeConfig Create(DbType? type = null, int? size = null, int? precision = null, int? scale = null, bool? isNullable = null)
+        public ITypeConfig Create(DbType? type = null, int? size = null, byte? precision = null, byte? scale = null, bool? isNullable = null)
         {
             if (!type.HasValue)
             {
